backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
supabase==2.9.0
aiosmtplib==3.0.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
email-validator==2.2.0
aiofiles==24.1.0

